app/api/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
pymupdf
markitdown[all]
beautifulsoup4
exa-py
anthropic
cloudflare
pydantic
mangum
